ux: move intersection and union instances on Constraint object
