+++ main/src/io/github/iltotore/iron/Constraint.scala
@@ -1,5 +1,8 @@
 package io.github.iltotore.iron
 
+import io.github.iltotore.iron.macros.union.*
+import io.github.iltotore.iron.macros.intersection.*
+
 /**
  * A typeclass representing the implementation of a constraint of a certain type of value.
  *
@@ -11,3 +14,22 @@ trait Constraint[A, C]:
   inline def test(value: A): Boolean
 
   inline def message: String
+
+object Constraint:
+  class UnionConstraint[A, C] extends Constraint[A, C]:
+
+    override inline def test(value: A): Boolean = unionCond[A, C](value)
+
+    override inline def message: String = unionMessage[A, C]
+
+  inline given [A, C](using inline u: IsUnion[C]): UnionConstraint[A, C] = new UnionConstraint
+
+  transparent inline given [C1, C2](using IsUnion[C1]): (C1 ==> C2) = unionImplication[C1, C2]
+
+  class IntersectionConstraint[A, C] extends Constraint[A, C]:
+
+    override inline def test(value: A): Boolean = intersectionCond[A, C](value)
+
+    override inline def message: String = intersectionMessage[A, C]
+
+  inline given [A, C](using inline i: IsIntersection[C]): IntersectionConstraint[A, C] = new IntersectionConstraint
